Move Design Guidance pages into ControlPages folder, code cleanup, add license headers
