+++ WinUIGallery/ControlPages/DesignGuidance/AccessibilityColorContrastPage.xaml
@@ -1,4 +1,16 @@
-<core:ItemsPageBase
+<!--
+//*********************************************************
+//
+// Copyright (c) Microsoft. All rights reserved.
+// THIS CODE IS PROVIDED *AS IS* WITHOUT WARRANTY OF
+// ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING ANY
+// IMPLIED WARRANTIES OF FITNESS FOR A PARTICULAR
+// PURPOSE, MERCHANTABILITY, OR NON-INFRINGEMENT.
+//
+//*********************************************************
+-->
+
+<Page
     x:Class="AppUIBasics.ControlPages.AccessibilityColorContrastPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
@@ -11,7 +23,7 @@
     Background="{ThemeResource HomePageBackgroundBrush}"
     mc:Ignorable="d">
 
-    <core:ItemsPageBase.Resources>
+    <Page.Resources>
         <Style x:Key="SubHeaderTextStyle" TargetType="TextBlock">
             <Setter Property="Margin" Value="0,24,0,0" />
             <Setter Property="FontSize" Value="16" />
@@ -25,7 +37,7 @@
         <Style TargetType="TextBlock">
             <Setter Property="TextWrapping" Value="WrapWholeWords" />
         </Style>
-    </core:ItemsPageBase.Resources>
+    </Page.Resources>
 
     <StackPanel
         MaxWidth="900"
@@ -291,4 +303,4 @@
             </Grid>
         </Grid>
     </StackPanel>
-</core:ItemsPageBase>
+</Page>
