+++ WinUIGallery/ControlPages/DesignGuidance/AccessibilityColorContrastPage.xaml.cs
@@ -1,18 +1,23 @@
+//*********************************************************
+//
+// Copyright (c) Microsoft. All rights reserved.
+// THIS CODE IS PROVIDED *AS IS* WITHOUT WARRANTY OF
+// ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING ANY
+// IMPLIED WARRANTIES OF FITNESS FOR A PARTICULAR
+// PURPOSE, MERCHANTABILITY, OR NON-INFRINGEMENT.
+//
+//*********************************************************
+
 using System;
-using AppUIBasics;
 using Microsoft.UI;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Media;
-using Microsoft.UI.Xaml.Navigation;
 using Microsoft.UI.Xaml.Shapes;
 using Windows.UI;
 
-// To learn more about WinUI, the WinUI project structure,
-// and more about our project templates, see: http://aka.ms/winui-project-info.
-
 namespace AppUIBasics.ControlPages
 {
-    public sealed partial class AccessibilityColorContrastPage : ItemsPageBase
+    public sealed partial class AccessibilityColorContrastPage : Page
     {
         public AccessibilityColorContrastPage()
         {
@@ -48,7 +53,7 @@ namespace AppUIBasics.ControlPages
             }
         }
 
-        public SolidColorBrush GetSolidColorBrush(string hex)
+        public static SolidColorBrush GetSolidColorBrush(string hex)
         {
             hex = hex.Replace("#", string.Empty);
             byte r = (byte)(Convert.ToUInt32(hex.Substring(0, 2), 16));
