+++ WinUIGallery/ControlPages/DesignGuidance/AccessibilityKeyboardPage.xaml
@@ -1,4 +1,15 @@
-<core:ItemsPageBase
+<!--
+//*********************************************************
+//
+// Copyright (c) Microsoft. All rights reserved.
+// THIS CODE IS PROVIDED *AS IS* WITHOUT WARRANTY OF
+// ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING ANY
+// IMPLIED WARRANTIES OF FITNESS FOR A PARTICULAR
+// PURPOSE, MERCHANTABILITY, OR NON-INFRINGEMENT.
+//
+//*********************************************************
+-->
+<Page
     x:Class="AppUIBasics.ControlPages.AccessibilityKeyboardPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
@@ -8,7 +19,7 @@
     Background="{ThemeResource HomePageBackgroundBrush}"
     mc:Ignorable="d">
 
-    <core:ItemsPageBase.Resources>
+    <Page.Resources>
         <Style x:Key="SubHeaderTextStyle" TargetType="TextBlock">
             <Setter Property="Margin" Value="0,24,0,0" />
             <Setter Property="FontSize" Value="16" />
@@ -22,7 +33,7 @@
         <Style TargetType="TextBlock">
             <Setter Property="TextWrapping" Value="WrapWholeWords" />
         </Style>
-    </core:ItemsPageBase.Resources>
+    </Page.Resources>
 
     <StackPanel
         MaxWidth="900"
@@ -366,4 +377,4 @@
             </core:ControlExample.Xaml>
         </core:ControlExample>
     </StackPanel>
-</core:ItemsPageBase>
+</Page>
