+++ WinUIGallery/ControlPages/DesignGuidance/AccessibilityScreenReaderPage.xaml
@@ -1,4 +1,16 @@
-<core:ItemsPageBase
+<!--
+//*********************************************************
+//
+// Copyright (c) Microsoft. All rights reserved.
+// THIS CODE IS PROVIDED *AS IS* WITHOUT WARRANTY OF
+// ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING ANY
+// IMPLIED WARRANTIES OF FITNESS FOR A PARTICULAR
+// PURPOSE, MERCHANTABILITY, OR NON-INFRINGEMENT.
+//
+//*********************************************************
+-->
+
+<Page
     x:Class="AppUIBasics.ControlPages.AccessibilityScreenReaderPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
@@ -10,7 +22,7 @@
     Background="{ThemeResource HomePageBackgroundBrush}"
     mc:Ignorable="d">
 
-    <core:ItemsPageBase.Resources>
+    <Page.Resources>
         <Style x:Key="SubHeaderTextStyle" TargetType="TextBlock">
             <Setter Property="Margin" Value="0,24,0,0" />
             <Setter Property="FontSize" Value="16" />
@@ -19,7 +31,7 @@
         <Style x:Key="SampleExplanationTextStyle" TargetType="TextBlock">
             <Setter Property="FontStyle" Value="Italic" />
         </Style>
-    </core:ItemsPageBase.Resources>
+    </Page.Resources>
 
     <StackPanel
         MaxWidth="900"
@@ -165,4 +177,4 @@ AutomationProperties.Name="Contacts"&gt;
             </core:ControlExample.Xaml>
         </core:ControlExample>
     </StackPanel>
-</core:ItemsPageBase>
+</Page>
