+++ WinUIGallery/ControlPages/DesignGuidance/AccessibilityScreenReaderPage.xaml.cs
@@ -1,0 +1,22 @@
+//*********************************************************
+//
+// Copyright (c) Microsoft. All rights reserved.
+// THIS CODE IS PROVIDED *AS IS* WITHOUT WARRANTY OF
+// ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING ANY
+// IMPLIED WARRANTIES OF FITNESS FOR A PARTICULAR
+// PURPOSE, MERCHANTABILITY, OR NON-INFRINGEMENT.
+//
+//*********************************************************
+
+using Microsoft.UI.Xaml.Controls;
+
+namespace AppUIBasics.ControlPages
+{
+    public sealed partial class AccessibilityScreenReaderPage : Page
+    {
+        public AccessibilityScreenReaderPage()
+        {
+            this.InitializeComponent();
+        }
+    }
+}
