+++ WinUIGallery/ControlPages/DesignGuidance/IconsPage.xaml
@@ -1,10 +1,21 @@
-<core:ItemsPageBase
+<!--
+//*********************************************************
+//
+// Copyright (c) Microsoft. All rights reserved.
+// THIS CODE IS PROVIDED *AS IS* WITHOUT WARRANTY OF
+// ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING ANY
+// IMPLIED WARRANTIES OF FITNESS FOR A PARTICULAR
+// PURPOSE, MERCHANTABILITY, OR NON-INFRINGEMENT.
+//
+//*********************************************************
+-->
+
+<Page
     x:Class="AppUIBasics.ControlPages.IconsPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:controls="using:AppUIBasics.Controls"
     xmlns:controls1="using:WinUIGallery.DesktopWap.Controls"
-    xmlns:core="using:AppUIBasics"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     Background="{ThemeResource HomePageBackgroundBrush}"
@@ -312,4 +323,4 @@
             </Expander>
         </Grid>
     </StackPanel>
-</core:ItemsPageBase>
+</Page>
