--- WinUIGallery/Controls/DesignGuidance/AccessibilityScreenReaderPage.xaml.cs
@@ -1,11 +1,0 @@
-
-namespace AppUIBasics.ControlPages
-{
-    public sealed partial class AccessibilityScreenReaderPage : ItemsPageBase
-    {
-        public AccessibilityScreenReaderPage()
-        {
-            this.InitializeComponent();
-        }
-    }
-}
